+++ ServerNAC/src/quartoSir/nac/cgt/bo/ProdutoBO.java
@@ -8,9 +8,7 @@ import org.hibernate.Session;
 import org.hibernate.Transaction;
 
 import quartoSir.nac.cgd.GenericDAO;
-import quartoSir.nac.domain.Cliente;
 import quartoSir.nac.domain.ItemPedido;
-import quartoSir.nac.domain.Pedido;
 import quartoSir.nac.domain.Produto;
 import quartoSir.nac.domain.util.ReturnObject;
 
@@ -23,9 +21,9 @@ public class ProdutoBO {
 	}
 	
 	/**
-	 * Realiza as verificações e salva o produto
+	 * Realiza as verificacoes e salva o produto
 	 * @param prod
-	 * @return ReturnObject indica o status da transação e mensagens de erro se existir.
+	 * @return ReturnObject indica o status da transacao e mensagens de erro se existir.
 	 */
 	public ReturnObject salvaProduto(Produto prod){
 		ReturnObject retorno = new ReturnObject();
@@ -39,19 +37,19 @@ public class ProdutoBO {
 		try{
 			
 			if(prod.getDescricao() == null || prod.getDescricao().equals("")){
-				retorno.setMensagem("ERRO: PRODUTO não pode ser vazio");
+				retorno.setMensagem("ERRO: PRODUTO nao pode ser vazio");
 				t.rollback();
 			}else if(prod.getEstoque() == null){
-				retorno.setMensagem("ERRO: ESTOQUE não pode ser vazio");
+				retorno.setMensagem("ERRO: ESTOQUE nao pode ser vazio");
 				t.rollback();
 			}else if(prod.getEstoque() < 0){
-				retorno.setMensagem("ERRO: ESTOQUE não pode ser negativo");
+				retorno.setMensagem("ERRO: ESTOQUE nao pode ser negativo");
 				t.rollback();
 			}else if(prod.getId() == null && prod.getEstoque() < 10){
-				retorno.setMensagem("ERRO: ESTOQUE inicial não pode ser inferior a 10");
+				retorno.setMensagem("ERRO: ESTOQUE inicial nao pode ser inferior a 10");
 				t.rollback();
 			}else if(prod.getPreco() == null){
-				retorno.setMensagem("ERRO: PRECO não pode ser vazio");
+				retorno.setMensagem("ERRO: PRECO nao pode ser vazio");
 				t.rollback();
 			}else if(prod.getPreco().compareTo(decimalZero) == -1 || prod.getPreco().compareTo(decimalZero) == 0) {
 				retorno.setMensagem("ERRO: PRECO deve ser maior que zero.");
@@ -60,7 +58,7 @@ public class ProdutoBO {
 				dao.save(prod);
 				t.commit();
 				retorno.setSucesso(true);
-				retorno.setMensagem("Produto cadastrado com sucesso!");
+				retorno.setMensagem("> Produto cadastrado com sucesso!");
 			}
 			
 		}catch(Exception e){
@@ -73,7 +71,7 @@ public class ProdutoBO {
 	/**
 	 * Pesquisa o produto por ID
 	 * @param prod
-	 * @return ReturnList indica o retorno da pesquisa e o status da transação, além de mensagens de erro se existir.
+	 * @return ReturnList indica o retorno da pesquisa e o status da transacao, alem de mensagens de erro se existir.
 	 */
 	public ReturnObject pesquisaProdutoID(Produto prod){
 		ReturnObject retorno = new ReturnObject();
@@ -83,7 +81,7 @@ public class ProdutoBO {
 		
 		try{
 			
-			if(prod.getId() != null){
+			if(prod.getId() != null  && !prod.getId().equals("")){
 				
 				Produto resultado = (Produto) dao.getById(Produto.class, prod.getId());
 				
@@ -91,13 +89,13 @@ public class ProdutoBO {
 				if(resultado != null){
 					retorno.setObj(resultado);
 				}else{
-					retorno.setMensagem("Nenhum produto encontrado com esse ID");
+					retorno.setMensagem("> Nenhum produto encontrado com esse ID");
 				}
 				
 				t.commit();
 				
 			}else{
-				retorno.setMensagem("ERRO: ID não pode ser vazio para essa consulta");
+				retorno.setMensagem("ERRO: ID nao pode ser vazio para essa consulta");
 				t.rollback();
 			}
 			
@@ -112,7 +110,7 @@ public class ProdutoBO {
 	/**
 	 * Pesquisa o produto pela descricao
 	 * @param cli
-	 * @return ReturnList indica o retorno da pesquisa e o status da transação, além de mensagens de erro se existir.
+	 * @return ReturnList indica o retorno da pesquisa e o status da transacao, alem de mensagens de erro se existir.
 	 */
 	public ReturnObject pesquisaProdutoDescricao(Produto prod){
 		ReturnObject retorno = new ReturnObject();
@@ -132,12 +130,12 @@ public class ProdutoBO {
 				if(resultados != null && resultados.size() > 0){
 					retorno.setLista(resultados);
 				}else{
-					retorno.setMensagem("Nenhum cliente encontrado com esse NOME");
+					retorno.setMensagem("> Nenhum produto encontrado com essa DESCRICAO");
 				}
 				t.commit();
 				
 			}else{
-				retorno.setMensagem("ERRO: NOME não pode ser vazio para essa consulta");
+				retorno.setMensagem("ERRO: Descricao nao pode ser vazia para essa consulta");
 				t.rollback();
 			}
 			
@@ -153,31 +151,43 @@ public class ProdutoBO {
 	/**
 	 * Atualiza o produto.
 	 * @param prod
-	 * @return ReturnObject indica o status e a mensagem da operação 
+	 * @return ReturnObject indica o status e a mensagem da operacao 
 	 */
 	public ReturnObject atualizaProduto(Produto prod) {
 		
-		Session session = dao.getSession();
-		Transaction t = session.beginTransaction();
+		BigDecimal decimalZero = new BigDecimal(0);
 		
 		ReturnObject retorno = new ReturnObject();	
 		retorno.setSucesso(false);
 		
+		Session session = dao.getSession();
+		Transaction t = session.beginTransaction();
+		
 		try{
-			
-			ItemPedidoBO itemPedidoBO = new ItemPedidoBO();
-			List<ItemPedido> itensPedido = itemPedidoBO.listaItensPedidoProduto(prod);
-			
-			if(itensPedido != null && itensPedido.size() > 0){
-				Produto prodAntigo = (Produto) dao.getById(Produto.class, prod.getId());	
+			if(prod.getDescricao() == null || prod.getDescricao().equals("")){
+				retorno.setMensagem("ERRO: PRODUTO nao pode ser vazio");
+				t.rollback();
+			}else if(prod.getEstoque() == null){
+				retorno.setMensagem("ERRO: ESTOQUE nao pode ser vazio");
+				t.rollback();
+			}else if(prod.getEstoque() < 0){
+				retorno.setMensagem("ERRO: ESTOQUE nao pode ser negativo");
+				t.rollback();
+			}else if(prod.getId() == null && prod.getEstoque() < 10){
+				retorno.setMensagem("ERRO: ESTOQUE inicial nao pode ser inferior a 10");
+				t.rollback();
+			}else if(prod.getPreco() == null){
+				retorno.setMensagem("ERRO: PRECO nao pode ser vazio");
+				t.rollback();
+			}else if(prod.getPreco().compareTo(decimalZero) == -1 || prod.getPreco().compareTo(decimalZero) == 0) {
+				retorno.setMensagem("ERRO: PRECO deve ser maior que zero.");
+				t.rollback();
+			}else{
+				dao.update(prod);						
+				t.commit();		
+				retorno.setSucesso(true);
+				retorno.setMensagem("> Produto alterado");
 			}
-			
-			dao.update(prod);						
-			t.commit();
-					
-			retorno.setSucesso(true);
-			retorno.setMensagem("Produto alterado");
-			
 		}catch(Exception e){
 			t.rollback();
 		}
@@ -186,29 +196,41 @@ public class ProdutoBO {
 		
 	}
 	
-	
+	/**
+	 * Deleta o produto.
+	 * @param prod
+	 * @return ReturnObject indica o status e a mensagem da operacao
+	 */
 	public ReturnObject excluiProduto(Produto prod) {
-		
-		Session session = dao.getSession();
-		Transaction t = session.beginTransaction();
 		
 		ReturnObject retorno = new ReturnObject();	
 		retorno.setSucesso(false);
 		
+		ItemPedidoBO itemPedidoBO = new ItemPedidoBO();
+		List<ItemPedido> itensPedido = itemPedidoBO.listaItensPedidoProduto(prod);
+		
+		Session session = dao.getSession();
+		Transaction t = session.beginTransaction();
+		
 		try{
-			ItemPedidoBO itemPedidoBO = new ItemPedidoBO();
-			List<ItemPedido> itensPedido = itemPedidoBO.listaItensPedidoProduto(prod);
 			
-			if(itensPedido != null && itensPedido.size() > 0){
-				retorno.setMensagem("ERRO: Não é excluir um produto que já possui pedidos.");
-				t.rollback();
+			if(prod.getId() != null && !prod.getId().equals("")){
+				
+				if(itensPedido != null && itensPedido.size() > 0){
+					retorno.setMensagem("ERRO: nao excluir um produto que ja possui pedidos.");
+					t.rollback();
+				}else{
+					dao.delete(prod);
+					t.commit();
+					retorno.setSucesso(true);
+					retorno.setMensagem("> Produto excluido com sucesso!");
+				}
+				
 			}else{
-				dao.delete(prod);
-				t.commit();
-				retorno.setSucesso(true);
-				retorno.setMensagem("Produto excluido com sucesso!");
-			}
-			
+				retorno.setMensagem("ERRO: ID nao pode ser vazio para essa consulta");
+				t.rollback();
+			}	
+
 		}catch(Exception e){
 			t.rollback();
 		}
